refactor: 优化部分代码逻辑
